backend/requirements.txt
Django==4.2.9
Django==5.0.7
djangorestframework==3.14.0
djangorestframework==3.15.2
xmlschema==3.3.0
